# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TasteTracker API",
        "description": "API REST del restaurante para gestionar menú, pedidos, clientes y empleados.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/v1/menu": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Listar todos los productos del menú",
                "description": "Obtiene una lista completa del catálogo de menú, ordenada por nombre.",
                "operationId": "15b6e54bf7297bce4d05b3f6af0de21a",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Menu"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Crear un nuevo producto del menú",
                "description": "Crea un producto en el catálogo del menú.",
                "operationId": "06c9995aeb442009d7ee3b943de92f93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MenuCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/v1/menu/{menu}": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Obtener un producto del menú",
                "description": "Devuelve los detalles de un producto específico del menú.",
                "operationId": "3cebd43d080fadb57fcdc6e8c228c827",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "description": "ID del producto del menú",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Menu"
                ],
                "summary": "Actualizar un producto del menú",
                "description": "Actualiza campos de un producto del menú.",
                "operationId": "ddb93d3528703e45c214352433cfad0f",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "description": "ID del producto del menú",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MenuUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu"
                ],
                "summary": "Eliminar un producto del menú",
                "description": "Elimina un producto del catálogo de menú.",
                "operationId": "85327e4d1e4d06f2560b800b6aa90079",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "description": "ID del producto del menú",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Menu"
                ],
                "summary": "Actualizar parcialmente un producto del menú",
                "description": "Actualiza uno o más campos de un producto del menú.",
                "operationId": "5472142ae4a5b0b34fdfc6462dec88cc",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "description": "ID del producto del menú",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MenuUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Menu": {
                "required": [
                    "nombre",
                    "precio",
                    "categoria"
                ],
                "properties": {
                    "id_menu": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Hamburguesa Clásica"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 120
                    },
                    "categoria": {
                        "type": "string",
                        "example": "comidas"
                    },
                    "disponible": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "MenuCreate": {
                "required": [
                    "nombre",
                    "precio",
                    "categoria"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Té helado"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 25.5
                    },
                    "categoria": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "bebidas"
                    },
                    "disponible": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "MenuUpdate": {
                "properties": {
                    "nombre": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Té helado grande"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 29
                    },
                    "categoria": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "bebidas"
                    },
                    "disponible": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}